Added blurr and simple heart textures, added pause feature and pause screen in level scene, changed LevelHUD to LevelPauseHUD and added functionality, added HUD for game over screen, changed bestScore and time in WorldReader, as well as HP in LivingGameObject from float to int, added player hearts to level scene, implemented monitoring player health and updating aforementioned hearts, added MainMenuScene, which now serves as the go-to scene.

+++ src/com/qb/stompy/dataReaders/WorldReader.java
@@ -15,15 +15,15 @@ public class WorldReader {
     public static class WRLevel {
         protected WRVec2f size, position, characterPosition, backgroundTextureSize;
         protected String textureName;
-        protected float bestScore, time;
+        protected int bestScore, time;
         protected List<WRBlock> blocks;
         public WRVec2f getSize() {return size;}
         public WRVec2f getPosition() {return position;}
         public WRVec2f getCharacterPosition() {return characterPosition;}
         public WRVec2f getBackgroundTextureSize() {return backgroundTextureSize;}
         public String getTextureName() {return textureName;}
-        public float getBestScore() {return bestScore;}
-        public float getTime() {return time;}
+        public int getBestScore() {return bestScore;}
+        public int getTime() {return time;}
         public List<WRBlock> getBlocks() {return blocks;}
     }
